requirements.txt
fastapi
uvicorn
requests
google-cloud-firestore
python-multipart
